Move dfLogics.cpp and logicTables.cpp to a new namespace called TFE_ExternalData
